+++ src/ArchiveFormats/ExeFile.h
@@ -1,74 +1,72 @@
 #pragma once
 
-#include <cstdint>
 #include <string>
 #include <map>
-
-#include "../zero_templates.h" // IWYU pragma: keep
+#include "util/int.h"
 
 struct PeHeader {
-	uint32_t mMagic; // PE\0\0
-	uint16_t mMachine;
-	uint16_t mNumberOfSections;
-	uint32_t mTimeDateStamp;
-	uint32_t mPointerToSymbolTable;
-	uint32_t mNumberOfSymbols;
-	uint16_t mSizeOfOptionalHeader;
-	uint16_t mCharacteristics;
+	u32 mMagic; // PE\0\0
+	u16 mMachine;
+	u16 mNumberOfSections;
+	u32 mTimeDateStamp;
+	u32 mPointerToSymbolTable;
+	u32 mNumberOfSymbols;
+	u16 mSizeOfOptionalHeader;
+	u16 mCharacteristics;
 };
 
 struct Pe32OptionalHeader {
-	uint16_t mMagic; // 0x010b - PE32, 0x020b - PE32+ (64 bit)
-	uint8_t  mMajorLinkerVersion;
-	uint8_t  mMinorLinkerVersion;
-	uint32_t mSizeOfCode;
-	uint32_t mSizeOfInitializedData;
-	uint32_t mSizeOfUninitializedData;
-	uint32_t mAddressOfEntryPoint;
-	uint32_t mBaseOfCode;
-	uint32_t mBaseOfData;
-	uint32_t mImageBase;
-	uint32_t mSectionAlignment;
-	uint32_t mFileAlignment;
-	uint16_t mMajorOperatingSystemVersion;
-	uint16_t mMinorOperatingSystemVersion;
-	uint16_t mMajorImageVersion;
-	uint16_t mMinorImageVersion;
-	uint16_t mMajorSubsystemVersion;
-	uint16_t mMinorSubsystemVersion;
-	uint32_t mWin32VersionValue;
-	uint32_t mSizeOfImage;
-	uint32_t mSizeOfHeaders;
-	uint32_t mCheckSum;
-	uint16_t mSubsystem;
-	uint16_t mDllCharacteristics;
-	uint32_t mSizeOfStackReserve;
-	uint32_t mSizeOfStackCommit;
-	uint32_t mSizeOfHeapReserve;
-	uint32_t mSizeOfHeapCommit;
-	uint32_t mLoaderFlags;
-	uint32_t mNumberOfRvaAndSizes;
+	u16 mMagic; // 0x010b - PE32, 0x020b - PE32+ (64 bit)
+	u8 mMajorLinkerVersion;
+	u8 mMinorLinkerVersion;
+	u32 mSizeOfCode;
+	u32 mSizeOfInitializedData;
+	u32 mSizeOfUninitializedData;
+	u32 mAddressOfEntryPoint;
+	u32 mBaseOfCode;
+	u32 mBaseOfData;
+	u32 mImageBase;
+	u32 mSectionAlignment;
+	u32 mFileAlignment;
+	u16 mMajorOperatingSystemVersion;
+	u16 mMinorOperatingSystemVersion;
+	u16 mMajorImageVersion;
+	u16 mMinorImageVersion;
+	u16 mMajorSubsystemVersion;
+	u16 mMinorSubsystemVersion;
+	u32 mWin32VersionValue;
+	u32 mSizeOfImage;
+	u32 mSizeOfHeaders;
+	u32 mCheckSum;
+	u16 mSubsystem;
+	u16 mDllCharacteristics;
+	u32 mSizeOfStackReserve;
+	u32 mSizeOfStackCommit;
+	u32 mSizeOfHeapReserve;
+	u32 mSizeOfHeapCommit;
+	u32 mLoaderFlags;
+	u32 mNumberOfRvaAndSizes;
 };
 
 struct Pe32SectionHeader {
 	char name[8];
-	uint32_t virtualSize;
-	uint32_t virtualAddress;
-	uint64_t sizeOfRawData;
-	uint32_t pointerToRawData;
-	uint32_t pointerToRelocations;
-	uint32_t pointerToLinenumbers;
-	uint16_t numberOfRelocations;
-	uint16_t numberOfLinenumbers;
-	uint32_t characteristics;
+	u32 virtualSize;
+	u32 virtualAddress;
+	u32 sizeOfRawData;
+	u32 pointerToRawData;
+	u32 pointerToRelocations;
+	u32 pointerToLinenumbers;
+	u16 numberOfRelocations;
+	u16 numberOfLinenumbers;
+	u32 characteristics;
 };
 
 class ExeFile {
     public:
-		uint8_t *buffer;
+		u8 *buffer;
 		PeHeader header;
 		std::map<std::string, Pe32SectionHeader> sections;
-		ExeFile(uint8_t *buffer) {
+		ExeFile(u8 *buffer) {
 			this->buffer = buffer;
 			this->header = GetPEHeader();
 			this->sections = ParseSectionHeaders();
@@ -83,5 +81,5 @@ class ExeFile {
 		Pe32SectionHeader* GetSectionHeader(std::string target_section);
 		std::map<std::string, Pe32SectionHeader> ParseSectionHeaders();
 
-        static bool SigCheck(uint8_t *buffer);
+        static bool SigCheck(u8 *buffer);
 };
